Create hashCracker.py
has a function to create hashes of either strings or bytes with a hasher of the user's choice (default is sha-256).  also has a function that brute forces the original string, given its hash.  if it is salted, it will return the original string with the salt on it.  also if it has more than 7 characters, and the list of characters it is given is not shorter than the default, it will have to look through an effectively infinite number of possible combinations.