Create heapIntArray.c
takes two command-line arguments.  allocates an amount of memory that is the first argument.  then it prompts the user for the integers.  then it re-allocates memory for an array the size of the second argument.  if the new value is larger, it will prompt for extra values, otherwise, it will copy over values up to the length of the new array.  after this, it will display the array to the standard output in the same form that python would.