Create lang.js
The language is a"Compiled" and interpreted language, written in ASCII codes, either decimal or hexadecimal, and has the option to have sections of it be interpreted directly as JavaScript.  The code that is not directly interpreted as JavaScript gets interpreted character by character and changed from ASCII codes to actual characters.  It also supports single-lined comments and has a function that changes JavaScript into hexadecimal that the interpreter and compiler can understand.  It is also very funny to show the programs to other people.