Minor Bug Fixes
When compiling to hex, it would compile "\0", "\t", "\n", and "\r" incorrectly.  this is because it took a decimal number and converted it to hexadecimal, and when it did this, it left off the extra zero at the beginning that the compiler and interpreter require to be there to work.  so it would compile "\r" into 0xa instead of 0x0a etc.