Refactor and Rename lang.js
Errors are now reported at the exact location they happen, instead of where in the grid of tokens the token was.  There is a new feature to download the code after it is compiled, as well as new classes for intrinsic keywords and supported characters.  Compiling from JavaScript to HexaScript is now simpler as you can combine multiple scripts with one function call.  The file has been renamed to HexaScript.js, which is subject to change.