Update HexaScript.js
change "js(" and "js)" to just "(" and ")", because the old way was dumb.  Introduced lbl, goto, push, print, if, then, stop, skip, pop, and shift.  There is a problem with goto, which is the same as in ti basic, that if you use goto to leave a block, it causes a memory leak. Although it doesn't matter all that much other than a console warning upon script finish.  The supported characters are now expanded, but at this point, I don't even remember what the point of the class of the supported characters was.