Create VerbalMemoryCheatsV2.js
Version 2. The last version was a separate html page and was manually used, which left room for errors. (../HTML/VerbalMemoryCheats.html)
This one is all javascript and mean to be used in the console, one the same page.
in the minified versions, the for loops only end if they get to exactly zero, so they will never end if the input isn't a positive integer.