Update PasswordGenerator.py
uses secrets.choice instead of random.choice, because it is more secure. also now it shuffles the characters beforehand and shuffles the password after creating int.
if the file is the main file, it also pauses at the end so you can copy the password before the window closes, and uses input for the parameters.