add fstrdiff to library
fstrdiff returns the index of the first difference between strings
if either is not a string, it returns NaN. It is considered a string if it is the built-in string or if
(the input).__type__() returns string. if there is no difference it returns -1.
I also added a comment at the end with a formula for pi.