update HelloWorldFunny class to be more in line with Java practices
you can't forget the interface and the abstract class.
Oh, and NEVER leave out the manager or the manager factory.
Java is such a disgusting language; I hate it so much.
It is literally the bane of my existence.
Also, no ar because if AP Computer Science A is still using Java 8, then obviously so should I.
It doesn't matter that it is like a decade old, it is a perfect language in a perfect state.