add EBSCO and JSTOR pdf rip files.
they rip PDFs. You still need to have access to the files, but normally you have to download the PDFSs like 1 page at a time, and it is usually limited to like 15 pages per day. I don't want to download a book slowly over a month. I want it now. Getting the service was not free, I paid for the books, so fucking give them to me.