update C/roslyn/
Uses literal string concatenation instead of %s for the compile-time constant strings; for some reason, GCC doesn't optimize it out (I checked with the mingw32-w64 devkit `strings`). The LANG macro is now supposed to be either cs or vb, where previously it was csc or vbc. The ROOT macro is no longer expected to have a quote at the start of the string.