update C/exe-wrap
Add a second version that doesn't use the C standard library (kind of).
So it should have a smaller output executable size.
Change `strlen(FILE_PATH)` to `sizeof(FILE_PATH) - 1` so it is more obvious that it is a compile-time operation.
for whereis[23]?.c, the version 2 is the default. use `-Dv1` to use version 1.