add C/scramble.c
Scrambles a buffer, probably not in a secure or fast way.
I don't recommend anyone use this ever.
The buffer string input should be hex-encoded, and the output is printed hex-encoded.
If compiled with `-DNO_MAIN`, it will not have a main function, and can be used elsewhere.
Use `void scramble(uint8_t *const buffer, size_t n);` as you function prototype if this is the case.
If `-DDEBUG` is given, the main function will print out the before and after buffers.
If `-DNO_MAIN` is given, `-DDEBUG` is ignored.