add PowerShell/invert-svg.ps1
invert the colors in an SVG.
it doesn't work with things like gradiants, because it only changed `fill` and `stroke` colors.
if there is no background rectangle, it creates one, and by default sets it to white (before inversion),
   this is because my purpose for this is for PDF inversion, although I haven't made sure it actually works for that.
It can probably be improved to use `StringBuilder` from .NET, but I don't feel like it.
It supports rgb, rgba, hsl, hsla, all types of hex codes (with and without the #), and all 147 named colors.
It uses the input file as the output file, since that is what I want, and I have school tomorrow so I don't care enough to change it