add ContinuedRadical.wl
I think it is interesting that the continued radical sequences for things like e and \pi are not in the OEIS as of today.
also, I don't know if continued radicals are a common thing, I kind of just made up the name, so there might be a different name already for what I am doing.