update the Java hello world with OOP
it is now arount 10 times the size. removing the javadoc from the new one de-bloats the file by 3.4 or 4.1 times, depending on if you use lines or characters as your measurement. I also added some extra classes and stuff, but it is so convoluted that you won't find them if you try.