add `NASM/api+crt/cat.nasm`
basically a clone of the `cat` Linux utility, but it doesn't work with the pipeline, it exits immediately when given no arguments, it prints different error messages, and it prints the error messages in red. It also doesn't use a buffer, and it just reads a byte and writes a byte in a loop.