optimize buffer size for `adler32.nasm` with a performance testing script
also change the buffer size from 1 MiB to 64 KiB.
I left in the testing script in case the optimal size is different on different architectures or if partial loop unrolling changes it or something.
To decide the best size, I just did `./adler32-perftest-bufsize --tex-only`, copy pasted the result into Desmos, and eyeballed where the drop off was.